requirements.txt
flask
flask-mysql
mysql-connector-python
